fix: mf-2482 handling profile card on right edge
