+++ packages/mask/src/social-network-adaptor/twitter.com/injection/ProfileCard/constants.ts
@@ -1,0 +1,2 @@
+export const CARD_WIDTH = 450
+export const CARD_HEIGHT = 500
