+++ packages/mask/src/social-network-adaptor/twitter.com/injection/ProfileCard/index.tsx
@@ -2,12 +2,14 @@ import { CrossIsolationMessages, ProfileIdentifier } from '@masknet/shared-base'
 import { LoadingBase, makeStyles } from '@masknet/theme'
 import { Twitter } from '@masknet/web3-providers'
 import type { SocialIdentity } from '@masknet/web3-shared-base'
-import { CSSProperties, useCallback, useEffect, useRef, useState } from 'react'
+import { useEffect, useRef, useState } from 'react'
 import { useAsync } from 'react-use'
 import { useSocialIdentity } from '../../../../components/DataSource/useActivatedUI.js'
 import { ProfileCard } from '../../../../components/InjectedComponents/ProfileCard/index.js'
 import { createReactRootShadowed } from '../../../../utils/index.js'
 import { twitterBase } from '../../base.js'
+import { CARD_HEIGHT, CARD_WIDTH } from './constants.js'
+import { useControlProfileCard } from './useControlProfileCard.js'
 
 export function injectProfileCardHolder(signal: AbortSignal) {
     const root = document.createElement('div')
@@ -19,8 +21,6 @@ export function injectProfileCardHolder(signal: AbortSignal) {
     createReactRootShadowed(root.shadowRoot, { signal }).render(<ProfileCardHolder />)
 }
 
-const CARD_WIDTH = 450
-const CARD_HEIGHT = 500
 const useStyles = makeStyles()((theme) => ({
     root: {
         position: 'absolute',
@@ -39,102 +39,17 @@ const useStyles = makeStyles()((theme) => ({
 
 function ProfileCardHolder() {
     const { classes } = useStyles()
-    const [style, setStyle] = useState<CSSProperties>({
-        visibility: 'hidden',
-        borderRadius: 10,
-    })
-    const activeRef = useRef(false)
     const holderRef = useRef<HTMLDivElement>(null)
-    const closeTimerRef = useRef<NodeJS.Timeout>()
     const [twitterId, setTwitterId] = useState('')
 
-    const hideProfileCard = useCallback(() => {
-        if (activeRef.current) return
-        setStyle((old) => {
-            if (old.visibility === 'hidden') return old
-            return {
-                ...old,
-                visibility: 'hidden',
-            }
-        })
-    }, [])
-
-    const showProfileCard = useCallback((patchStyle: CSSProperties) => {
-        clearTimeout(closeTimerRef.current)
-        setStyle((old) => {
-            const { visibility, left, top } = old
-            if (visibility === 'visible' && left === patchStyle.left && top === patchStyle.top) return old
-            return { ...old, ...patchStyle, visibility: 'visible' }
-        })
-    }, [])
-
-    useEffect(() => {
-        const holder = holderRef.current
-        if (!holder) return
-        const enter = () => {
-            activeRef.current = true
-            clearTimeout(closeTimerRef.current)
-        }
-        const leave = () => {
-            activeRef.current = false
-            clearTimeout(closeTimerRef.current)
-            closeTimerRef.current = setTimeout(hideProfileCard, 2000)
-        }
-        holder.addEventListener('mouseenter', enter)
-        holder.addEventListener('mouseleave', leave)
-        return () => {
-            holder.removeEventListener('mouseenter', enter)
-            holder.removeEventListener('mouseleave', leave)
-        }
-    }, [hideProfileCard])
+    const style = useControlProfileCard(holderRef)
 
     useEffect(() => {
         return CrossIsolationMessages.events.profileCardEvent.on((event) => {
-            if (!event.open) {
-                hideProfileCard()
-                return
-            }
-            const { userId, badgeBounding: bounding } = event
-            setTwitterId(userId)
-            const reachedBottomBoundary = bounding.top + bounding.height + CARD_HEIGHT > window.innerHeight
-            let x = Math.max(bounding.left + bounding.width / 2 - CARD_WIDTH / 2, 0)
-            let y = bounding.top + bounding.height
-            if (reachedBottomBoundary) {
-                const reachedTopBoundary = bounding.top < CARD_HEIGHT
-                if (reachedTopBoundary) {
-                    x = bounding.left + bounding.width
-                    y = Math.min(window.innerHeight - CARD_HEIGHT, Math.max(bounding.top - CARD_HEIGHT / 2))
-                } else {
-                    y = bounding.top - CARD_HEIGHT
-                }
-            }
-            // Prefer to show top left corner of the card.
-            x = Math.max(0, x)
-            y = Math.max(0, y)
-
-            const pageOffset = document.scrollingElement?.scrollTop || 0
-            const newLeft = x
-            const newTop = y + pageOffset
-            showProfileCard({
-                left: newLeft,
-                top: newTop,
-            })
+            if (!event.open) return
+            setTwitterId(event.userId)
         })
-    }, [hideProfileCard, showProfileCard])
-
-    useEffect(() => {
-        const onClick = (event: MouseEvent) => {
-            // @ts-ignore
-            // `NODE.contains(other)` doesn't work for cross multiple layer of Shadow DOM
-            if (!event.path?.includes(holderRef.current)) {
-                hideProfileCard()
-            }
-        }
-        document.body.addEventListener('click', onClick)
-        return () => {
-            document.body.removeEventListener('click', onClick)
-        }
-    }, [hideProfileCard])
+    }, [])
 
     const { value: identity, loading } = useAsync(async (): Promise<SocialIdentity | null> => {
         if (!twitterId) return null
